fix: make all devtools load on import
