+++ fission/package.json
@@ -22,6 +22,7 @@
     "style": "bun run fmt && bun run lint",
     "style:fix": "bun run fmt:fix && bun run lint:fix",
     "assetpack": "git lfs pull && tar -xf public/assetpack.zip -C public/",
+    "assetpack:update": "cd public && zip -FS -r assetpack.zip Downloadables",
     "playwright:install": "bun x playwright install",
     "electron:start": "electron-forge start",
     "electron:package": "electron-forge package",
