+++ fission/src/ui/panels/DeveloperToolPanel.tsx
@@ -5,87 +5,12 @@ import MirabufCachingService, { MiraType } from "@/mirabuf/MirabufLoader"
 import MirabufSceneObject from "@/mirabuf/MirabufSceneObject"
 import { mirabuf } from "@/proto/mirabuf"
 import PreferencesSystem from "@/systems/preferences/PreferencesSystem"
-import {
-    defaultFieldPreferences,
-    type FieldPreferences,
-    type ScoringZonePreferences,
-} from "@/systems/preferences/PreferenceTypes"
 import World from "@/systems/World"
-import FieldMiraEditor, { type DevtoolMiraData } from "../../mirabuf/FieldMiraEditor"
+import FieldMiraEditor, { type DevtoolKey, devtoolHandlers, devtoolKeys } from "../../mirabuf/FieldMiraEditor"
 import { globalAddToast } from "../components/GlobalUIControls"
 import type { PanelImplProps } from "../components/Panel"
 import { LabelWithTooltip } from "../components/StyledComponents"
 import { useUIContext } from "../helpers/UIProviderHelpers"
-
-const devtoolHandlers = {
-    "devtool:scoring_zones": {
-        get(field) {
-            return field.fieldPreferences?.scoringZones ?? defaultFieldPreferences().scoringZones
-        },
-        set(field, val) {
-            val ??= defaultFieldPreferences().scoringZones
-            if (!field.fieldPreferences || !this.validate(val)) return
-            field.fieldPreferences.scoringZones = val
-            field.updateScoringZones()
-        },
-        validate(val): val is ScoringZonePreferences[] {
-            if (!Array.isArray(val)) return false
-            return val.every(
-                z =>
-                    typeof z === "object" &&
-                    z !== null &&
-                    typeof z.name === "string" &&
-                    (z.alliance === "red" || z.alliance === "blue") &&
-                    (typeof z.parentNode === "string" || z.parentNode === undefined) &&
-                    typeof z.points === "number" &&
-                    typeof z.destroyGamepiece === "boolean" &&
-                    typeof z.persistentPoints === "boolean" &&
-                    Array.isArray(z.deltaTransformation)
-            )
-        },
-    },
-    "devtool:spawn_locations": {
-        get(field) {
-            return field.fieldPreferences?.spawnLocations ?? defaultFieldPreferences().spawnLocations
-        },
-        set(field, val) {
-            val ??= defaultFieldPreferences().spawnLocations
-            if (!field.fieldPreferences || !this.validate(val)) return
-            field.fieldPreferences.spawnLocations = val
-        },
-        validate(val: unknown): val is FieldPreferences["spawnLocations"] {
-            const isStructureCorrect =
-                typeof val === "object" && val != null && "red" in val && "blue" in val && "default" in val
-
-            if (!isStructureCorrect) return false
-            return (["red", "blue"] as const).every(v => {
-                const obj = val[v]
-                if (!(typeof obj === "object" && obj != null && 1 in obj && 2 in obj && 3 in obj)) return false
-                return ([1, 2, 3] as const).every(v => {
-                    const spawnposition = obj[v]
-                    return (
-                        typeof spawnposition == "object" &&
-                        spawnposition != null &&
-                        "pos" in spawnposition &&
-                        "yaw" in spawnposition &&
-                        Array.isArray(spawnposition["pos"]) &&
-                        spawnposition["pos"].length == 3 &&
-                        typeof spawnposition["yaw"] == "number"
-                    )
-                })
-            })
-        },
-    },
-} as const satisfies Partial<{
-    [K in keyof DevtoolMiraData]: {
-        get(field: MirabufSceneObject): DevtoolMiraData[K]
-        set(field: MirabufSceneObject, val: unknown | null): void
-        validate(val: unknown): val is DevtoolMiraData[K]
-    }
-}>
-
-type DevtoolKey = keyof typeof devtoolHandlers
-const DEVTOOL_KEYS = Object.keys(devtoolHandlers) as DevtoolKey[]
 
 function getCurrentFieldObj() {
     for (const obj of World.sceneRenderer.sceneObjects.values()) {
@@ -323,16 +248,18 @@ const DeveloperToolPanel: React.FC<PanelImplProps<void, void>> = ({ panel }) => 
                         </ul>
                         <div className="mt-2 border-t border-gray-600 pt-2">
                             <div className="text-xs mb-1 text-gray-300">Add new:</div>
-                            {DEVTOOL_KEYS.filter(k => !keys.includes(k)).map(key => (
-                                <Button
-                                    key={key}
-                                    onClick={() => handleAdd(key)}
-                                    className="w-full mb-1 whitespace-normal break-words"
-                                >
-                                    {key}
-                                </Button>
-                            ))}
-                            {DEVTOOL_KEYS.filter(k => !keys.includes(k)).length === 0 && (
+                            {devtoolKeys
+                                .filter(k => !keys.includes(k))
+                                .map(key => (
+                                    <Button
+                                        key={key}
+                                        onClick={() => handleAdd(key)}
+                                        className="w-full mb-1 whitespace-normal break-words"
+                                    >
+                                        {key}
+                                    </Button>
+                                ))}
+                            {devtoolKeys.filter(k => !keys.includes(k)).length === 0 && (
                                 <div className="text-gray-400 italic text-xs">All keys added</div>
                             )}
                         </div>
